+++ src/daplis/functions/calibrate.py
@@ -34,6 +34,8 @@ functions:
 
 """
 
+from __future__ import annotations
+
 import glob
 import os
 import sys