+++ src/daplis/functions/cross_talk.py
@@ -29,6 +29,8 @@ functions:
 
 """
 
+from __future__ import annotations
+
 import glob
 import os
 import pickle