+++ src/daplis/functions/fits.py
@@ -31,6 +31,8 @@ functions:
 
 """
 
+from __future__ import annotations
+
 import glob
 import os
 import pickle