+++ src/daplis/functions/unpack.py
@@ -13,6 +13,8 @@ functions:
 
 """
 
+from __future__ import annotations
+
 import os
 
 import numpy as np